Download all available dates

Lists all available dates for a MODIS Land Products Subset product at a particular location.

list_dates(product = NULL, lat = NULL, lon = NULL, site_id = NULL)

Arguments

product
a valid MODIS product name

lat
latitude in decimal degrees

lon
longitude in decimal degrees

site_id
site id (overides lat / lon)

Value

A data frame of all available dates for a MODIS Land Products Subsets products at the given location.

Examples

# list all available MODIS Land Products Subsets products
bands <- list_dates(product = "MOD11A2", lat = 40, lon = -110)
print(bands)

#>     modis_date calendar_date
#> 1     A2000049    2000-02-18
#> 2     A2000057    2000-02-26
#> 3     A2000065    2000-03-05
#> 4     A2000073    2000-03-13
#> 5     A2000081    2000-03-21
#> 6     A2000089    2000-03-29
#> 7     A2000097    2000-04-06
#> 8     A2000105    2000-04-14
#> 9     A2000113    2000-04-22
#> 10    A2000121    2000-04-30
#> 11    A2000129    2000-05-08
#> 12    A2000137    2000-05-16
#> 13    A2000145    2000-05-24
#> 14    A2000153    2000-06-01
#> 15    A2000161    2000-06-09
#> 16    A2000169    2000-06-17
#> 17    A2000177    2000-06-25
#> 18    A2000185    2000-07-03
#> 19    A2000193    2000-07-11
#> 20    A2000201    2000-07-19
#> 21    A2000209    2000-07-27
#> 22    A2000217    2000-08-04
#> 23    A2000225    2000-08-12
#> 24    A2000233    2000-08-20
#> 25    A2000241    2000-08-28
#> 26    A2000249    2000-09-05
#> 27    A2000257    2000-09-13
#> 28    A2000265    2000-09-21
#> 29    A2000273    2000-09-29
#> 30    A2000281    2000-10-07
#> 31    A2000289    2000-10-15
#> 32    A2000297    2000-10-23
#> 33    A2000305    2000-10-31
#> 34    A2000313    2000-11-08
#> 35    A2000321    2000-11-16
#> 36    A2000329    2000-11-24
#> 37    A2000337    2000-12-02
#> 38    A2000345    2000-12-10
#> 39    A2000353    2000-12-18
#> 40    A2000361    2000-12-26
#> 41    A2001001    2001-01-01
#> 42    A2001009    2001-01-09
#> 43    A2001017    2001-01-17
#> 44    A2001025    2001-01-25
#> 45    A2001033    2001-02-02
#> 46    A2001041    2001-02-10
#> 47    A2001049    2001-02-18
#> 48    A2001057    2001-02-26
#> 49    A2001065    2001-03-06
#> 50    A2001073    2001-03-14
#> 51    A2001081    2001-03-22
#> 52    A2001089    2001-03-30
#> 53    A2001097    2001-04-07
#> 54    A2001105    2001-04-15
#> 55    A2001113    2001-04-23
#> 56    A2001121    2001-05-01
#> 57    A2001129    2001-05-09
#> 58    A2001137    2001-05-17
#> 59    A2001145    2001-05-25
#> 60    A2001153    2001-06-02
#> 61    A2001161    2001-06-10
#> 62    A2001177    2001-06-26
#> 63    A2001185    2001-07-04
#> 64    A2001193    2001-07-12
#> 65    A2001201    2001-07-20
#> 66    A2001209    2001-07-28
#> 67    A2001217    2001-08-05
#> 68    A2001225    2001-08-13
#> 69    A2001233    2001-08-21
#> 70    A2001241    2001-08-29
#> 71    A2001249    2001-09-06
#> 72    A2001257    2001-09-14
#> 73    A2001265    2001-09-22
#> 74    A2001273    2001-09-30
#> 75    A2001281    2001-10-08
#> 76    A2001289    2001-10-16
#> 77    A2001297    2001-10-24
#> 78    A2001305    2001-11-01
#> 79    A2001313    2001-11-09
#> 80    A2001321    2001-11-17
#> 81    A2001329    2001-11-25
#> 82    A2001337    2001-12-03
#> 83    A2001345    2001-12-11
#> 84    A2001353    2001-12-19
#> 85    A2001361    2001-12-27
#> 86    A2002001    2002-01-01
#> 87    A2002009    2002-01-09
#> 88    A2002017    2002-01-17
#> 89    A2002025    2002-01-25
#> 90    A2002033    2002-02-02
#> 91    A2002041    2002-02-10
#> 92    A2002049    2002-02-18
#> 93    A2002057    2002-02-26
#> 94    A2002065    2002-03-06
#> 95    A2002073    2002-03-14
#> 96    A2002081    2002-03-22
#> 97    A2002089    2002-03-30
#> 98    A2002097    2002-04-07
#> 99    A2002105    2002-04-15
#> 100   A2002113    2002-04-23
#> 101   A2002121    2002-05-01
#> 102   A2002129    2002-05-09
#> 103   A2002137    2002-05-17
#> 104   A2002145    2002-05-25
#> 105   A2002153    2002-06-02
#> 106   A2002161    2002-06-10
#> 107   A2002169    2002-06-18
#> 108   A2002177    2002-06-26
#> 109   A2002185    2002-07-04
#> 110   A2002193    2002-07-12
#> 111   A2002201    2002-07-20
#> 112   A2002209    2002-07-28
#> 113   A2002217    2002-08-05
#> 114   A2002225    2002-08-13
#> 115   A2002233    2002-08-21
#> 116   A2002241    2002-08-29
#> 117   A2002249    2002-09-06
#> 118   A2002257    2002-09-14
#> 119   A2002265    2002-09-22
#> 120   A2002273    2002-09-30
#> 121   A2002281    2002-10-08
#> 122   A2002289    2002-10-16
#> 123   A2002297    2002-10-24
#> 124   A2002305    2002-11-01
#> 125   A2002313    2002-11-09
#> 126   A2002321    2002-11-17
#> 127   A2002329    2002-11-25
#> 128   A2002337    2002-12-03
#> 129   A2002345    2002-12-11
#> 130   A2002353    2002-12-19
#> 131   A2002361    2002-12-27
#> 132   A2003001    2003-01-01
#> 133   A2003009    2003-01-09
#> 134   A2003017    2003-01-17
#> 135   A2003025    2003-01-25
#> 136   A2003033    2003-02-02
#> 137   A2003041    2003-02-10
#> 138   A2003049    2003-02-18
#> 139   A2003057    2003-02-26
#> 140   A2003065    2003-03-06
#> 141   A2003073    2003-03-14
#> 142   A2003081    2003-03-22
#> 143   A2003089    2003-03-30
#> 144   A2003097    2003-04-07
#> 145   A2003105    2003-04-15
#> 146   A2003113    2003-04-23
#> 147   A2003121    2003-05-01
#> 148   A2003129    2003-05-09
#> 149   A2003137    2003-05-17
#> 150   A2003145    2003-05-25
#> 151   A2003153    2003-06-02
#> 152   A2003161    2003-06-10
#> 153   A2003169    2003-06-18
#> 154   A2003177    2003-06-26
#> 155   A2003185    2003-07-04
#> 156   A2003193    2003-07-12
#> 157   A2003201    2003-07-20
#> 158   A2003209    2003-07-28
#> 159   A2003217    2003-08-05
#> 160   A2003225    2003-08-13
#> 161   A2003233    2003-08-21
#> 162   A2003241    2003-08-29
#> 163   A2003249    2003-09-06
#> 164   A2003257    2003-09-14
#> 165   A2003265    2003-09-22
#> 166   A2003273    2003-09-30
#> 167   A2003281    2003-10-08
#> 168   A2003289    2003-10-16
#> 169   A2003297    2003-10-24
#> 170   A2003305    2003-11-01
#> 171   A2003313    2003-11-09
#> 172   A2003321    2003-11-17
#> 173   A2003329    2003-11-25
#> 174   A2003337    2003-12-03
#> 175   A2003345    2003-12-11
#> 176   A2003353    2003-12-19
#> 177   A2003361    2003-12-27
#> 178   A2004001    2004-01-01
#> 179   A2004009    2004-01-09
#> 180   A2004017    2004-01-17
#> 181   A2004025    2004-01-25
#> 182   A2004033    2004-02-02
#> 183   A2004041    2004-02-10
#> 184   A2004049    2004-02-18
#> 185   A2004057    2004-02-26
#> 186   A2004065    2004-03-05
#> 187   A2004073    2004-03-13
#> 188   A2004081    2004-03-21
#> 189   A2004089    2004-03-29
#> 190   A2004097    2004-04-06
#> 191   A2004105    2004-04-14
#> 192   A2004113    2004-04-22
#> 193   A2004121    2004-04-30
#> 194   A2004129    2004-05-08
#> 195   A2004137    2004-05-16
#> 196   A2004145    2004-05-24
#> 197   A2004153    2004-06-01
#> 198   A2004161    2004-06-09
#> 199   A2004169    2004-06-17
#> 200   A2004177    2004-06-25
#> 201   A2004185    2004-07-03
#> 202   A2004193    2004-07-11
#> 203   A2004201    2004-07-19
#> 204   A2004209    2004-07-27
#> 205   A2004217    2004-08-04
#> 206   A2004225    2004-08-12
#> 207   A2004233    2004-08-20
#> 208   A2004241    2004-08-28
#> 209   A2004249    2004-09-05
#> 210   A2004257    2004-09-13
#> 211   A2004265    2004-09-21
#> 212   A2004273    2004-09-29
#> 213   A2004281    2004-10-07
#> 214   A2004289    2004-10-15
#> 215   A2004297    2004-10-23
#> 216   A2004305    2004-10-31
#> 217   A2004313    2004-11-08
#> 218   A2004321    2004-11-16
#> 219   A2004329    2004-11-24
#> 220   A2004337    2004-12-02
#> 221   A2004345    2004-12-10
#> 222   A2004353    2004-12-18
#> 223   A2004361    2004-12-26
#> 224   A2005001    2005-01-01
#> 225   A2005009    2005-01-09
#> 226   A2005017    2005-01-17
#> 227   A2005025    2005-01-25
#> 228   A2005033    2005-02-02
#> 229   A2005041    2005-02-10
#> 230   A2005049    2005-02-18
#> 231   A2005057    2005-02-26
#> 232   A2005065    2005-03-06
#> 233   A2005073    2005-03-14
#> 234   A2005081    2005-03-22
#> 235   A2005089    2005-03-30
#> 236   A2005097    2005-04-07
#> 237   A2005105    2005-04-15
#> 238   A2005113    2005-04-23
#> 239   A2005121    2005-05-01
#> 240   A2005129    2005-05-09
#> 241   A2005137    2005-05-17
#> 242   A2005145    2005-05-25
#> 243   A2005153    2005-06-02
#> 244   A2005161    2005-06-10
#> 245   A2005169    2005-06-18
#> 246   A2005177    2005-06-26
#> 247   A2005185    2005-07-04
#> 248   A2005193    2005-07-12
#> 249   A2005201    2005-07-20
#> 250   A2005209    2005-07-28
#> 251   A2005217    2005-08-05
#> 252   A2005225    2005-08-13
#> 253   A2005233    2005-08-21
#> 254   A2005241    2005-08-29
#> 255   A2005249    2005-09-06
#> 256   A2005257    2005-09-14
#> 257   A2005265    2005-09-22
#> 258   A2005273    2005-09-30
#> 259   A2005281    2005-10-08
#> 260   A2005289    2005-10-16
#> 261   A2005297    2005-10-24
#> 262   A2005305    2005-11-01
#> 263   A2005313    2005-11-09
#> 264   A2005321    2005-11-17
#> 265   A2005329    2005-11-25
#> 266   A2005337    2005-12-03
#> 267   A2005345    2005-12-11
#> 268   A2005353    2005-12-19
#> 269   A2005361    2005-12-27
#> 270   A2006001    2006-01-01
#> 271   A2006009    2006-01-09
#> 272   A2006017    2006-01-17
#> 273   A2006025    2006-01-25
#> 274   A2006033    2006-02-02
#> 275   A2006041    2006-02-10
#> 276   A2006049    2006-02-18
#> 277   A2006057    2006-02-26
#> 278   A2006065    2006-03-06
#> 279   A2006073    2006-03-14
#> 280   A2006081    2006-03-22
#> 281   A2006089    2006-03-30
#> 282   A2006097    2006-04-07
#> 283   A2006105    2006-04-15
#> 284   A2006113    2006-04-23
#> 285   A2006121    2006-05-01
#> 286   A2006129    2006-05-09
#> 287   A2006137    2006-05-17
#> 288   A2006145    2006-05-25
#> 289   A2006153    2006-06-02
#> 290   A2006161    2006-06-10
#> 291   A2006169    2006-06-18
#> 292   A2006177    2006-06-26
#> 293   A2006185    2006-07-04
#> 294   A2006193    2006-07-12
#> 295   A2006201    2006-07-20
#> 296   A2006209    2006-07-28
#> 297   A2006217    2006-08-05
#> 298   A2006225    2006-08-13
#> 299   A2006233    2006-08-21
#> 300   A2006241    2006-08-29
#> 301   A2006249    2006-09-06
#> 302   A2006257    2006-09-14
#> 303   A2006265    2006-09-22
#> 304   A2006273    2006-09-30
#> 305   A2006281    2006-10-08
#> 306   A2006289    2006-10-16
#> 307   A2006297    2006-10-24
#> 308   A2006305    2006-11-01
#> 309   A2006313    2006-11-09
#> 310   A2006321    2006-11-17
#> 311   A2006329    2006-11-25
#> 312   A2006337    2006-12-03
#> 313   A2006345    2006-12-11
#> 314   A2006353    2006-12-19
#> 315   A2006361    2006-12-27
#> 316   A2007001    2007-01-01
#> 317   A2007009    2007-01-09
#> 318   A2007017    2007-01-17
#> 319   A2007025    2007-01-25
#> 320   A2007033    2007-02-02
#> 321   A2007041    2007-02-10
#> 322   A2007049    2007-02-18
#> 323   A2007057    2007-02-26
#> 324   A2007065    2007-03-06
#> 325   A2007073    2007-03-14
#> 326   A2007081    2007-03-22
#> 327   A2007089    2007-03-30
#> 328   A2007097    2007-04-07
#> 329   A2007105    2007-04-15
#> 330   A2007113    2007-04-23
#> 331   A2007121    2007-05-01
#> 332   A2007129    2007-05-09
#> 333   A2007137    2007-05-17
#> 334   A2007145    2007-05-25
#> 335   A2007153    2007-06-02
#> 336   A2007161    2007-06-10
#> 337   A2007169    2007-06-18
#> 338   A2007177    2007-06-26
#> 339   A2007185    2007-07-04
#> 340   A2007193    2007-07-12
#> 341   A2007201    2007-07-20
#> 342   A2007209    2007-07-28
#> 343   A2007217    2007-08-05
#> 344   A2007225    2007-08-13
#> 345   A2007233    2007-08-21
#> 346   A2007241    2007-08-29
#> 347   A2007249    2007-09-06
#> 348   A2007257    2007-09-14
#> 349   A2007265    2007-09-22
#> 350   A2007273    2007-09-30
#> 351   A2007281    2007-10-08
#> 352   A2007289    2007-10-16
#> 353   A2007297    2007-10-24
#> 354   A2007305    2007-11-01
#> 355   A2007313    2007-11-09
#> 356   A2007321    2007-11-17
#> 357   A2007329    2007-11-25
#> 358   A2007337    2007-12-03
#> 359   A2007345    2007-12-11
#> 360   A2007353    2007-12-19
#> 361   A2007361    2007-12-27
#> 362   A2008001    2008-01-01
#> 363   A2008009    2008-01-09
#> 364   A2008017    2008-01-17
#> 365   A2008025    2008-01-25
#> 366   A2008033    2008-02-02
#> 367   A2008041    2008-02-10
#> 368   A2008049    2008-02-18
#> 369   A2008057    2008-02-26
#> 370   A2008065    2008-03-05
#> 371   A2008073    2008-03-13
#> 372   A2008081    2008-03-21
#> 373   A2008089    2008-03-29
#> 374   A2008097    2008-04-06
#> 375   A2008105    2008-04-14
#> 376   A2008113    2008-04-22
#> 377   A2008121    2008-04-30
#> 378   A2008129    2008-05-08
#> 379   A2008137    2008-05-16
#> 380   A2008145    2008-05-24
#> 381   A2008153    2008-06-01
#> 382   A2008161    2008-06-09
#> 383   A2008169    2008-06-17
#> 384   A2008177    2008-06-25
#> 385   A2008185    2008-07-03
#> 386   A2008193    2008-07-11
#> 387   A2008201    2008-07-19
#> 388   A2008209    2008-07-27
#> 389   A2008217    2008-08-04
#> 390   A2008225    2008-08-12
#> 391   A2008233    2008-08-20
#> 392   A2008241    2008-08-28
#> 393   A2008249    2008-09-05
#> 394   A2008257    2008-09-13
#> 395   A2008265    2008-09-21
#> 396   A2008273    2008-09-29
#> 397   A2008281    2008-10-07
#> 398   A2008289    2008-10-15
#> 399   A2008297    2008-10-23
#> 400   A2008305    2008-10-31
#> 401   A2008313    2008-11-08
#> 402   A2008321    2008-11-16
#> 403   A2008329    2008-11-24
#> 404   A2008337    2008-12-02
#> 405   A2008345    2008-12-10
#> 406   A2008353    2008-12-18
#> 407   A2008361    2008-12-26
#> 408   A2009001    2009-01-01
#> 409   A2009009    2009-01-09
#> 410   A2009017    2009-01-17
#> 411   A2009025    2009-01-25
#> 412   A2009033    2009-02-02
#> 413   A2009041    2009-02-10
#> 414   A2009049    2009-02-18
#> 415   A2009057    2009-02-26
#> 416   A2009065    2009-03-06
#> 417   A2009073    2009-03-14
#> 418   A2009081    2009-03-22
#> 419   A2009089    2009-03-30
#> 420   A2009097    2009-04-07
#> 421   A2009105    2009-04-15
#> 422   A2009113    2009-04-23
#> 423   A2009121    2009-05-01
#> 424   A2009129    2009-05-09
#> 425   A2009137    2009-05-17
#> 426   A2009145    2009-05-25
#> 427   A2009153    2009-06-02
#> 428   A2009161    2009-06-10
#> 429   A2009169    2009-06-18
#> 430   A2009177    2009-06-26
#> 431   A2009185    2009-07-04
#> 432   A2009193    2009-07-12
#> 433   A2009201    2009-07-20
#> 434   A2009209    2009-07-28
#> 435   A2009217    2009-08-05
#> 436   A2009225    2009-08-13
#> 437   A2009233    2009-08-21
#> 438   A2009241    2009-08-29
#> 439   A2009249    2009-09-06
#> 440   A2009257    2009-09-14
#> 441   A2009265    2009-09-22
#> 442   A2009273    2009-09-30
#> 443   A2009281    2009-10-08
#> 444   A2009289    2009-10-16
#> 445   A2009297    2009-10-24
#> 446   A2009305    2009-11-01
#> 447   A2009313    2009-11-09
#> 448   A2009321    2009-11-17
#> 449   A2009329    2009-11-25
#> 450   A2009337    2009-12-03
#> 451   A2009345    2009-12-11
#> 452   A2009353    2009-12-19
#> 453   A2009361    2009-12-27
#> 454   A2010001    2010-01-01
#> 455   A2010009    2010-01-09
#> 456   A2010017    2010-01-17
#> 457   A2010025    2010-01-25
#> 458   A2010033    2010-02-02
#> 459   A2010041    2010-02-10
#> 460   A2010049    2010-02-18
#> 461   A2010057    2010-02-26
#> 462   A2010065    2010-03-06
#> 463   A2010073    2010-03-14
#> 464   A2010081    2010-03-22
#> 465   A2010089    2010-03-30
#> 466   A2010097    2010-04-07
#> 467   A2010105    2010-04-15
#> 468   A2010113    2010-04-23
#> 469   A2010121    2010-05-01
#> 470   A2010129    2010-05-09
#> 471   A2010137    2010-05-17
#> 472   A2010145    2010-05-25
#> 473   A2010153    2010-06-02
#> 474   A2010161    2010-06-10
#> 475   A2010169    2010-06-18
#> 476   A2010177    2010-06-26
#> 477   A2010185    2010-07-04
#> 478   A2010193    2010-07-12
#> 479   A2010201    2010-07-20
#> 480   A2010209    2010-07-28
#> 481   A2010217    2010-08-05
#> 482   A2010225    2010-08-13
#> 483   A2010233    2010-08-21
#> 484   A2010241    2010-08-29
#> 485   A2010249    2010-09-06
#> 486   A2010257    2010-09-14
#> 487   A2010265    2010-09-22
#> 488   A2010273    2010-09-30
#> 489   A2010281    2010-10-08
#> 490   A2010289    2010-10-16
#> 491   A2010297    2010-10-24
#> 492   A2010305    2010-11-01
#> 493   A2010313    2010-11-09
#> 494   A2010321    2010-11-17
#> 495   A2010329    2010-11-25
#> 496   A2010337    2010-12-03
#> 497   A2010345    2010-12-11
#> 498   A2010353    2010-12-19
#> 499   A2010361    2010-12-27
#> 500   A2011001    2011-01-01
#>  [ reached getOption("max.print") -- omitted 338 rows ]