Download all available dates

Lists all available dates for a MODIS Land Products Subset product at a particular location.

mt_dates(product, lat, lon, site_id, network)

Arguments

product
a valid MODIS product name

lat
latitude in decimal degrees

lon
longitude in decimal degrees

site_id
site id (overides lat / lon)

network
the network for which to generate the site list, when not provided the complete list is provided

Value

A data frame of all available dates for a MODIS Land Products Subsets products at the given location.

See also

Examples

# \donttest{
# list all available MODIS Land Products Subsets products
bands <- mt_dates(product = "MOD11A2", lat = 40, lon = -110)

#> Request failed [500]. Retrying in 1 seconds...

#> Request failed [500]. Retrying in 3.3 seconds...

#> Error: 7b0a20202274797065223a202261626f75743a626c616e6b222c0a2020227469746c65223a2022496e7465726e616c20536572766572204572726f72222c0a20202264657461696c223a20225468652073657276657220656e636f756e746572656420616e20696e7465726e616c206572726f7220616e642077617320756e61626c6520746f20636f6d706c65746520796f757220726571756573742e204569746865722074686520736572766572206973206f7665726c6f61646564206f7220746865726520697320616e206572726f7220696e20746865206170706c69636174696f6e2e222c0a202022737461747573223a203530300a7d0a

head(bands)

#> Error in head(bands): object 'bands' not found

# }