Read a MODISTools data file

Read a MODISTools data file into a nested structure of class MODISTools.

mt_read(filename)

Arguments

filename
a MODISTools data file

Value

A nested data structure including site meta-data, the full header and the data as a `data.frame()`.

Examples

# download data
mt_subset(product = "MOD11A2",
           lat = 40,
           lon = -110,
           band = "LST_Day_1km",
           start = "2004-01-01",
           end = "2004-03-31",
           internal = FALSE)

# read file
df <- mt_read(paste0(tempdir(),
               "/sitename_MOD11A2_2004-01-01_2004-03-31.csv"))
print(df)

#> $header
#> $header$xllcorner
#> [1] "-9370036.39"
#> 
#> $header$yllcorner
#> [1] "4447802.08"
#> 
#> $header$cellsize
#> [1] "926.625433055834"
#> 
#> $header$nrows
#> [1] 1
#> 
#> $header$ncols
#> [1] 1
#> 
#> $header$band
#> [1] "LST_Day_1km"
#> 
#> $header$units
#> [1] "Kelvin"
#> 
#> $header$scale
#> [1] "0.02"
#> 
#> $header$latitude
#> [1] 40
#> 
#> $header$longitude
#> [1] -110
#> 
#> $header$site
#> [1] "sitename"
#> 
#> $header$product
#> [1] "MOD11A2"
#> 
#> $header$start
#> [1] "2004-01-01"
#> 
#> $header$end
#> [1] "2004-03-31"
#> 
#> $header$complete
#> [1] TRUE
#> 
#> 
#> $data
#>    modis_date calendar_date        band   tile     proc_date pixel  data
#> 1    A2004001    2004-01-01 LST_Day_1km h09v05 2015212185706     1 13098
#> 2    A2004009    2004-01-09 LST_Day_1km h09v05 2015212201022     1 13062
#> 3    A2004017    2004-01-17 LST_Day_1km h09v05 2015212213103     1 13297
#> 4    A2004025    2004-01-25 LST_Day_1km h09v05 2015213005429     1 13323
#> 5    A2004033    2004-02-02 LST_Day_1km h09v05 2015213090158     1 13315
#> 6    A2004041    2004-02-10 LST_Day_1km h09v05 2015213165253     1 13227
#> 7    A2004049    2004-02-18 LST_Day_1km h09v05 2015216182840     1 13739
#> 8    A2004057    2004-02-26 LST_Day_1km h09v05 2015217020635     1 13783
#> 9    A2004065    2004-03-05 LST_Day_1km h09v05 2015217091316     1 14748
#> 10   A2004073    2004-03-13 LST_Day_1km h09v05 2015217164857     1 15105
#> 11   A2004081    2004-03-21 LST_Day_1km h09v05 2015217230030     1 15290
#> 12   A2004089    2004-03-29 LST_Day_1km h09v05 2015218055617     1 15410
#> 
#> attr(,"class")
#> [1] "MODISTools"