Download all available fixed sites

Lists all available MODIS Land Products Subset pre-processed sites

mt_sites(network)

Arguments

network

the network for which to generate the site list, when not provided the complete list is provided

Value

A data frame of all available MODIS Land Products Subsets pre-processed sites

See also

Examples

# \donttest{
# list all available MODIS Land Products Subsets products
sites <- mt_sites()
print(head(sites))
#>                   siteid                sitename network latitude longitude
#> 1  ae_abudhabi_abu_dhabi            AE Abu Dhabi AERONET 24.47611  54.32889
#> 2     ae_abudhabi_al_ain     AE Abu Dhabi Al Ain AERONET 24.24217  55.70517
#> 3  ae_abudhabi_al_dhafra  AE Abu Dhabi Al Dhafra AERONET 24.25428  54.54967
#> 4 ae_abudhabi_al_khaznah AE Abu Dhabi Al Khaznah AERONET 24.15861  55.10056
#> 5    ae_abudhabi_al_qlaa    AE Abu Dhabi Al Qlaa AERONET 24.13333  53.03333
#> 6      ae_abudhabi_dhabi            AE Abu Dhabi AERONET 24.48056  54.38278
#>      state              country
#> 1 AbuDhabi United Arab Emirates
#> 2 AbuDhabi United Arab Emirates
#> 3 AbuDhabi United Arab Emirates
#> 4 AbuDhabi United Arab Emirates
#> 5 AbuDhabi United Arab Emirates
#> 6 AbuDhabi United Arab Emirates
# }