Convert tidy MODISTools data to raster (stack)

Convert tidy MODISTools data to a raster (stack)

mt_to_raster(df, reproject = FALSE)

Arguments

df
a valid MODISTools data frame with a single band (filter for a particular band using the dplyr filter() function or base subset()

reproject
reproject output to lat / long (default = FALSE)

Value

A raster stack populated with the tidy dataframe values

See also

Examples

# \donttest{
# list all available MODIS Land Products Subsets products
# download data
LC <- mt_subset(product = "MCD12Q1",
 lat = 48.383662,
 lon = 2.610250,
 band = "LC_Type1",
 start = "2005-01-01",
 end = "2005-12-30",
 km_lr = 2,
 km_ab = 2,
 site_name = "testsite",
 internal = TRUE,
 progress = FALSE)

#> Error in open.connection(con, "rb") : HTTP error 500.

#> Error in (function (product) {    products <- MODISTools::mt_products()$product    if (missing(product) | !(product %in% products)) {        stop("please specify a product, or check your product name...")    }    url <- paste(mt_server(), product, "bands", sep = "/")    bands <- try(jsonlite::fromJSON(url))    if (inherits(bands, "try-error")) {        stop("Your requested timed out or the server is unreachable")    }    return(bands$bands)})(product = product): Your requested timed out or the server is unreachable

head(LC)

#> Error in head(LC): object 'LC' not found

# convert to raster
LC_r <- mt_to_raster(df = LC)

#> Error in is.data.frame(df): object 'LC' not found

# }