Convert tidy MODISTools data to terra SpatRaster

Convert tidy MODISTools data to a terra SpatRaster for easy spatial processing and plotting.

mt_to_terra(df, reproject = FALSE)

Arguments

df

a valid MODISTools data frame with a single band (filter for a particular band using the dplyr filter() function or base subset()

reproject

reproject output to lat / long (default = FALSE)

Value

A terra SpatRaster populated with the tidy dataframe values

See also

Examples

# \donttest{
# list all available MODIS Land Products Subsets products
# download data
LC <- mt_subset(product = "MCD12Q1",
 lat = 48.383662,
 lon = 2.610250,
 band = "LC_Type1",
 start = "2005-01-01",
 end = "2005-12-30",
 km_lr = 2,
 km_ab = 2,
 site_name = "testsite",
 internal = TRUE,
 progress = FALSE)

head(LC)
#>     xllcorner  yllcorner      cellsize nrows ncols     band units         scale
#> 1.1 190884.84 5378134.05 463.312716528     9     9 LC_Type1 class Not Available
#> 1.2 190884.84 5378134.05 463.312716528     9     9 LC_Type1 class Not Available
#> 1.3 190884.84 5378134.05 463.312716528     9     9 LC_Type1 class Not Available
#> 1.4 190884.84 5378134.05 463.312716528     9     9 LC_Type1 class Not Available
#> 1.5 190884.84 5378134.05 463.312716528     9     9 LC_Type1 class Not Available
#> 1.6 190884.84 5378134.05 463.312716528     9     9 LC_Type1 class Not Available
#>     latitude longitude     site product      start        end complete
#> 1.1 48.38366   2.61025 testsite MCD12Q1 2005-01-01 2005-12-30     TRUE
#> 1.2 48.38366   2.61025 testsite MCD12Q1 2005-01-01 2005-12-30     TRUE
#> 1.3 48.38366   2.61025 testsite MCD12Q1 2005-01-01 2005-12-30     TRUE
#> 1.4 48.38366   2.61025 testsite MCD12Q1 2005-01-01 2005-12-30     TRUE
#> 1.5 48.38366   2.61025 testsite MCD12Q1 2005-01-01 2005-12-30     TRUE
#> 1.6 48.38366   2.61025 testsite MCD12Q1 2005-01-01 2005-12-30     TRUE
#>     modis_date calendar_date   tile     proc_date pixel value
#> 1.1   A2005001    2005-01-01 h18v04 2022153213513     1     5
#> 1.2   A2005001    2005-01-01 h18v04 2022153213513     2     5
#> 1.3   A2005001    2005-01-01 h18v04 2022153213513     3     5
#> 1.4   A2005001    2005-01-01 h18v04 2022153213513     4     5
#> 1.5   A2005001    2005-01-01 h18v04 2022153213513     5     5
#> 1.6   A2005001    2005-01-01 h18v04 2022153213513     6     5

# convert to raster
LC_r <- mt_to_terra(df = LC)
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("character"), new("standardGeneric", .Data = function (x,     ...) standardGeneric("crs"), generic = structure("crs", package = "terra"),     package = "terra", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "crs"), domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘crs’ for signature ‘"character"’
# }