Convert sinusoidal coordinates to lat / lon

A full description of the sinusoidal projection is provided on the lpdaac page: https://lpdaac.usgs.gov/dataset_discovery/modis and wikipedia: https://en.wikipedia.org/wiki/Sinusoidal_projection

sin_to_ll(x, y)

Arguments

x

sinusoidal x coordinate (vector)

y

sinusoidal y coordinate (vector)

See also

Examples

# \donttest{
# Download some test data
subset <- mt_subset(product = "MOD11A2",
                        lat = 40,
                        lon = -110,
                        band = "LST_Day_1km",
                        start = "2004-01-01",
                        end = "2004-03-31",
                        progress = FALSE)

# convert sinusoidal to lat / lon
lat_lon <- sin_to_ll(subset$xllcorner, subset$yllcorner)
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("character"), new("standardGeneric", .Data = function (x,     ...) standardGeneric("crs"), generic = structure("crs", package = "terra"),     package = "terra", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "crs"), domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘crs’ for signature ‘"character"’

# bind with the original dataframe
subset <- cbind(subset, lat_lon)
#> Error in data.frame(..., check.names = FALSE): object 'lat_lon' not found
head(subset)
#>       xllcorner  yllcorner         cellsize nrows ncols        band  units
#> 1.1 -9370036.35 4446875.49 926.625433055834     1     1 LST_Day_1km Kelvin
#> 2.1 -9370036.35 4446875.49 926.625433055834     1     1 LST_Day_1km Kelvin
#> 3.1 -9370036.35 4446875.49 926.625433055834     1     1 LST_Day_1km Kelvin
#> 4.1 -9370036.35 4446875.49 926.625433055834     1     1 LST_Day_1km Kelvin
#> 5.1 -9370036.35 4446875.49 926.625433055834     1     1 LST_Day_1km Kelvin
#> 6.1 -9370036.35 4446875.49 926.625433055834     1     1 LST_Day_1km Kelvin
#>     scale latitude longitude     site product      start        end complete
#> 1.1  0.02       40      -110 sitename MOD11A2 2004-01-01 2004-03-31     TRUE
#> 2.1  0.02       40      -110 sitename MOD11A2 2004-01-01 2004-03-31     TRUE
#> 3.1  0.02       40      -110 sitename MOD11A2 2004-01-01 2004-03-31     TRUE
#> 4.1  0.02       40      -110 sitename MOD11A2 2004-01-01 2004-03-31     TRUE
#> 5.1  0.02       40      -110 sitename MOD11A2 2004-01-01 2004-03-31     TRUE
#> 6.1  0.02       40      -110 sitename MOD11A2 2004-01-01 2004-03-31     TRUE
#>     modis_date calendar_date   tile     proc_date pixel value
#> 1.1   A2004001    2004-01-01 h09v05 2020168005635     1 13129
#> 2.1   A2004009    2004-01-09 h09v05 2020168010833     1 13102
#> 3.1   A2004017    2004-01-17 h09v05 2020168012220     1 13343
#> 4.1   A2004025    2004-01-25 h09v05 2020168013617     1 13364
#> 5.1   A2004033    2004-02-02 h09v05 2020168015053     1 13364
#> 6.1   A2004041    2004-02-10 h09v05 2020168022045     1 13272
# }