Batch download MODIS Land Products subsets

Lists all available dates for a MODIS Land Products Subset product at a particular location.

mt_batch_subset(df = NULL, product = NULL, band = NULL,
  start = "2000-01-01", end = format(Sys.time(), "%Y-%m-%d"),
  km_lr = 0, km_ab = 0, out_dir = tempdir(), internal = TRUE,
  ncores = "auto")

Arguments

df
a file holding locations and their sitenames to batch process

product
a valid MODIS product name

band
band to download (default = NULL, all bands)

start
start date

end
end date

km_lr
km left-right to sample

km_ab
km above-below to sample

out_dir
location where to store all data

internal
should the data be returned as an internal data structure TRUE or FALSE (default = TRUE)

ncores
number of cores to use while downloading in parallel (auto will select the all cpu cores - 1)

Value

A nested list containing the downloaded data and a descriptive header with meta-data.

Examples

# create data frame with a site_name, lat and lon column
# holding the respective names of sites and their location
df <- data.frame("site_name" = paste("test",1:2))
df$lat <- 40
df$lon <- -110

print(df)

#>   site_name lat  lon
#> 1    test 1  40 -110
#> 2    test 2  40 -110

# test batch download
subsets <- mt_batch_subset(df = df,
                        product = "MOD11A2",
                        band = "LST_Day_1km",
                        internal = TRUE,
                        start = "2004-01-01",
                        end = "2004-03-31",
                        out_dir = "~")

print(str(subsets))

#> List of 2
#>  $ test 1:List of 2
#>   ..$ header:List of 15
#>   .. ..$ xllcorner: chr "-9370036.39"
#>   .. ..$ yllcorner: chr "4447802.08"
#>   .. ..$ cellsize : chr "926.625433055834"
#>   .. ..$ nrows    : int 1
#>   .. ..$ ncols    : int 1
#>   .. ..$ band     : chr "LST_Day_1km"
#>   .. ..$ units    : chr "Kelvin"
#>   .. ..$ scale    : chr "0.02"
#>   .. ..$ latitude : num 40
#>   .. ..$ longitude: num -110
#>   .. ..$ site     : chr "test 1"
#>   .. ..$ product  : chr "MOD11A2"
#>   .. ..$ start    : chr "2004-01-01"
#>   .. ..$ end      : chr "2004-03-31"
#>   .. ..$ complete : logi TRUE
#>   ..$ data  :'data.frame':	12 obs. of  7 variables:
#>   .. ..$ modis_date   : chr [1:12] "A2004001" "A2004009" "A2004017" "A2004025" ...
#>   .. ..$ calendar_date: chr [1:12] "2004-01-01" "2004-01-09" "2004-01-17" "2004-01-25" ...
#>   .. ..$ band         : chr [1:12] "LST_Day_1km" "LST_Day_1km" "LST_Day_1km" "LST_Day_1km" ...
#>   .. ..$ tile         : chr [1:12] "h09v05" "h09v05" "h09v05" "h09v05" ...
#>   .. ..$ proc_date    : chr [1:12] "2015212185706" "2015212201022" "2015212213103" "2015213005429" ...
#>   .. ..$ pixel        : chr [1:12] "1" "1" "1" "1" ...
#>   .. ..$ data         : int [1:12] 13098 13062 13297 13323 13315 13227 13739 13783 14748 15105 ...
#>   ..- attr(*, "class")= chr "MODISTools"
#>  $ test 2:List of 2
#>   ..$ header:List of 15
#>   .. ..$ xllcorner: chr "-9370036.39"
#>   .. ..$ yllcorner: chr "4447802.08"
#>   .. ..$ cellsize : chr "926.625433055834"
#>   .. ..$ nrows    : int 1
#>   .. ..$ ncols    : int 1
#>   .. ..$ band     : chr "LST_Day_1km"
#>   .. ..$ units    : chr "Kelvin"
#>   .. ..$ scale    : chr "0.02"
#>   .. ..$ latitude : num 40
#>   .. ..$ longitude: num -110
#>   .. ..$ site     : chr "test 2"
#>   .. ..$ product  : chr "MOD11A2"
#>   .. ..$ start    : chr "2004-01-01"
#>   .. ..$ end      : chr "2004-03-31"
#>   .. ..$ complete : logi TRUE
#>   ..$ data  :'data.frame':	12 obs. of  7 variables:
#>   .. ..$ modis_date   : chr [1:12] "A2004001" "A2004009" "A2004017" "A2004025" ...
#>   .. ..$ calendar_date: chr [1:12] "2004-01-01" "2004-01-09" "2004-01-17" "2004-01-25" ...
#>   .. ..$ band         : chr [1:12] "LST_Day_1km" "LST_Day_1km" "LST_Day_1km" "LST_Day_1km" ...
#>   .. ..$ tile         : chr [1:12] "h09v05" "h09v05" "h09v05" "h09v05" ...
#>   .. ..$ proc_date    : chr [1:12] "2015212185706" "2015212201022" "2015212213103" "2015213005429" ...
#>   .. ..$ pixel        : chr [1:12] "1" "1" "1" "1" ...
#>   .. ..$ data         : int [1:12] 13098 13062 13297 13323 13315 13227 13739 13783 14748 15105 ...
#>   ..- attr(*, "class")= chr "MODISTools"
#> NULL