Read a MODISTools data file

Read a MODISTools data file into a nested structure of class MODISTools.

mt_read(filename)

Arguments

filename
a MODISTools data file

Value

A nested data structure including site meta-data, the full header and the data as a `data.frame()`.

Examples

# download data
get_subset(product = "MOD11A2",
           lat = 40,
           lon = -110,
           band = "LST_Day_1km",
           start = "2004-01-01",
           end = "2004-03-31",
           internal = FALSE)

#> Error in get_subset(product = "MOD11A2", lat = 40, lon = -110, band = "LST_Day_1km",     start = "2004-01-01", end = "2004-03-31", internal = FALSE): could not find function "get_subset"

# read file
df <- read_subset(paste0(tempdir(),
               "/sitename_MOD11A2_2004-01-01_2004-03-31.csv"))

#> Error in read_subset(paste0(tempdir(), "/sitename_MOD11A2_2004-01-01_2004-03-31.csv")): could not find function "read_subset"

print(df)

#> function (x, df1, df2, ncp, log = FALSE) 
#> {
#>     if (missing(ncp)) 
#>         .Call(C_df, x, df1, df2, log)
#>     else .Call(C_dnf, x, df1, df2, ncp, log)
#> }
#> <bytecode: 0x5080e50>
#> <environment: namespace:stats>