Download MODIS Land Products subsets

Download a MODIS Land Products Subset product for a given point location buffered with a given amount of kilometers left-right, top-bottom for a given location (provided as latitude and longitude values).

mt_subset(
  product,
  band,
  lat,
  lon,
  start = "2000-01-01",
  end = format(Sys.time(), "%Y-%m-%d"),
  km_lr = 0,
  km_ab = 0,
  site_id,
  network,
  site_name = "sitename",
  out_dir = tempdir(),
  internal = TRUE,
  progress = TRUE
)

Arguments

product
a valid MODIS product name

band
band or bands (as a character vector) to download

lat
latitude in decimal degrees

lon
longitude in decimal degrees

start
start date

end
end date

km_lr
km left-right to sample (rounded to the nearest integer)

km_ab
km above-below to sample (rounded to the nearest integer)

site_id
site id (overides lat / lon)

network
the network for which to generate the site list, when not provided the complete list is provided

site_name
arbitrary site name used in writing data to file (default = sitename)

out_dir
path where to store the data if writing to disk (default = tempdir())

internal
should the data be returned as an internal data structure TRUE or FALSE (default = TRUE)

progress
show download progress

Value

A data frame combining meta-data and actual data values.

See also

Examples

# \donttest{
# list all available MODIS Land Products Subsets products
# download data
subset <- mt_subset(product = "MOD11A2",
                        lat = 40,
                        lon = -110,
                        band = "LST_Day_1km",
                        start = "2004-01-01",
                        end = "2004-03-31",
                        progress = FALSE)

#> Error in open.connection(con, "rb") : HTTP error 500.

#> Error in (function (product) {    products <- MODISTools::mt_products()$product    if (missing(product) | !(product %in% products)) {        stop("please specify a product, or check your product name...")    }    url <- paste(mt_server(), product, "bands", sep = "/")    bands <- try(jsonlite::fromJSON(url))    if (inherits(bands, "try-error")) {        stop("Your requested timed out or the server is unreachable")    }    return(bands$bands)})(product = product): Your requested timed out or the server is unreachable

 head(subset)

#>                      
#> 1 function (x, ...)  
#> 2 UseMethod("subset")

# }